+++ src/sorting/InsertionSort.java
@@ -5,10 +5,10 @@ import java.util.List;
 public class InsertionSort implements ISortAlgorithm{
 
     @Override
-    public List<Integer> sort(List<Integer> array) {
+    public List<Long> sort(List<Long> array) {
         int n = array.size();
         for (int i = 1; i < n; i++) {
-            int key = array.get(i);
+            Long key = array.get(i);
             int j = i - 1;
             while (j >= 0 && array.get(j) > key) {
                 array.set(j + 1, array.get(j));
